requirements.txt
streamlit
pinecone-client
torch
transformers
scikit-learn
Pillow
